Move user utils to utils folder
